Set up Manager and Employees associations, naive implementation of #not_managed_by
Set up self-joins for manager-employee associations.

Start working on the scope `#not_managed_by`, with a naive implementation.

We're correctly returning the employees who are not managed by the passed-in
manager, but according to the specs, we should also be returning all managers,
since managers have no manager and re therefore not managed by the passed in
manager.

The problem here is that we're using a `joins`, which does an `INNER JOIN` to
`managers`, which will by definition exclude records where `manager_id` is
`nil`.

UP NEXT: refactor to use `LEFT JOIN`.